Move logic to GpuOverrides
